continued
Its a start. Adjusted size on a couple photos and ID'd and adjusted width on a couple. Moved h3 above image so text is aligned. I don't love the fixed divs on grid when skinny, so will eventualy make a breakpoint and go to two images per row, and one for mobile.